+++ app/Models/TareaModel.php
@@ -36,13 +36,33 @@ class TareaModel extends \Com\Daw2\Core\BaseModel {
         return $usuarios;
     }
 
+    public function buscarTareaPorId(int $idTarea): ?array {
+        $stmt = $this->pdo->prepare("SELECT * FROM tareas ta JOIN proyectos p ON ta.id_proyecto=p.id_proyecto LEFT JOIN usuarios us "
+                . "ON ta.id_usuario_tarea_prop=us.id_usuario LEFT JOIN colores c ON ta.id_color_tarea = c.id_color "
+                . "WHERE id_tarea = ?");
+        $stmt->execute([$idTarea]);
+
+        $usuarioEncontrado = $stmt->fetch();
+
+        if ($usuarioEncontrado) {
+            return $usuarioEncontrado;
+        } else {
+            return null;
+        }
+    }
+
     public function contador(): int {
         $stmt = $this->pdo->query("SELECT COUNT(*) FROM tareas");
         return $stmt->fetchColumn();
     }
 
-    public function deleteTarea(int $idTarea): void {
-        $stmt = $this->pdo->prepare("DELETE FROM tareas WHERE id_tarea = ?");
-        $stmt->execute([$idTarea]);
+    public function deleteTarea(int $idTarea): bool {
+        if (!is_null($this->buscarTareaPorId($idTarea))) {
+            $stmt = $this->pdo->prepare("DELETE FROM tareas WHERE id_tarea = ?");
+            $stmt->execute([$idTarea]);
+            return true;
+        }
+
+        return false;
     }
 }
